accredible_connector/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2018.8.24        # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
click==6.7                # via flask
cryptography==2.3.1       # via pyopenssl, requests
dnspython==1.15.0         # via eventlet
eventlet==0.24.1
flask-wtf==0.14.2
flask==1.0.2
greenlet==0.4.15          # via eventlet
gunicorn==19.9.0
idna==2.7                 # via cryptography, requests
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
markupsafe==1.0           # via jinja2
monotonic==1.5            # via eventlet
pycparser==2.18           # via cffi
pyjwt==1.6.4
pyopenssl==18.0.0         # via requests
redis==2.10.6
requests[security]==2.19.1
six==1.11.0               # via cryptography, eventlet, pyopenssl
urllib3==1.23             # via requests
werkzeug==0.14.1          # via flask
wtforms==2.2.1            # via flask-wtf
